add MDenseRNN
